modified:   bin/AST/ASTBuilder.class 	modified:   bin/AST/ExprNodes/IntLiteralNode.class 	modified:   bin/AST/StmtNodes/DefStmtNode.class 	new file:   bin/IRNodes/binary.class 	new file:   bin/IRNodes/compare.class 	new file:   bin/IRNodes/constantValue.class 	new file:   bin/IRNodes/entity.class 	modified:   bin/SemanticChecker/SemanticChecker.class 	modified:   src/AST/ASTBuilder.java 	modified:   src/AST/ExprNodes/IntLiteralNode.java 	new file:   src/IR/IRTypes/IRClassType.java 	new file:   src/IR/IRTypes/IRPointerType.java 	new file:   src/IR/IRTypes/IRSimpleType.java 	new file:   src/IR/IRTypes/IRType.java 	new file:   src/IR/IRTypes/IRVoidType.java 	new file:   src/IR/entity/constantValue.java 	new file:   src/IR/entity/entity.java 	new file:   src/IR/entity/variable.java 	new file:   src/IR/inst/alloca.java 	new file:   src/IR/inst/binary.java 	new file:   src/IR/inst/block.java 	new file:   src/IR/inst/call.java 	new file:   src/IR/inst/compare.java 	new file:   src/IR/inst/conditionalBr.java 	new file:   src/IR/inst/getElementPtr.java 	new file:   src/IR/inst/inst.java 	new file:   src/IR/inst/load.java 	new file:   src/IR/inst/ret.java 	new file:   src/IR/inst/store.java 	new file:   src/IR/inst/unconditionalBr.java 	modified:   src/SemanticChecker/SemanticChecker.java 	modified:   test.mx

+++ src/AST/ASTBuilder.java
@@ -151,7 +151,7 @@ public class ASTBuilder extends MxBaseVisitor<ASTNode> {
 
     public ASTNode visitLiteral(MxParser.LiteralContext ctx) {
         if (ctx.DecimalInteger() != null) {
-            return new IntLiteralNode(Integer.parseInt(ctx.DecimalInteger().toString()), new position(ctx));
+            return new IntLiteralNode(Long.parseLong(ctx.DecimalInteger().toString()), new position(ctx));
         } else if (ctx.StringLiteral() != null) {
             String str = ctx.StringLiteral().toString();
             str = str.substring(1, str.length() - 1);
